+++ cpp/game/perfect_path_optimizer.h
@@ -23,7 +23,7 @@ class PerfectPathOptimizer : public PathOptimizerInterface {
   // Returns all possible decisions with associated score
   // score means the loss of time (lap-wise) compared to optimal choice
   // so atleast one lane will always have value of 0
-  std::map<Switch, int> Score(const Position& position) override;
+  std::map<Switch, double> Score(const Position& position) override;
 
   void Optimize(std::atomic<bool>* ready_flag);
